refactor(vm): rename `tx` module
