+++ crates/zksync/core/src/vm/factory_deps.rs
